AdjustedFreq {spopmodel}	R Documentation

Get adjusted frequency based on relative retention results.

Description

Provides a convenient way to adjust catch overall given results of gear selectivity (i.e., NetFet model results).

S3 method for class RelativeRetention.

Usage


AdjustedFreq(x, ...)

## Default S3 method:
AdjustedFreq(x, ...)

## S3 method for class 'RelativeRetention'
AdjustedFreq(x, ...)

Arguments

x
Generic for S3 methods.

...
Passed to other methods.

Value

A dataframe of catch and adjusted catch per length (row).

Examples


# code here

[Package spopmodel version 0.0.0.9000 Index]