ApplyNetFit {spopmodel}	R Documentation

Apply Millar's NetFit to catch data using gill net or trammel net.

Description

Apply Millar's NetFit to catch data using gill net or trammel net.

Usage


ApplyNetFit(data, len, mesh, meshUnit = "in", relPower = NULL)

Arguments

data
A dataframe with at least length and net mesh (size) fields.

len
Field name within data containing (numeric) length measurement.

mesh
Field name within data containing size of net mesh. Field data can be character but must be "coerceable" to numeric.

meshUnit
Scalar character containing unit of measure of net. Choices are "in" (inches), "ft" (feet), "mm" (millimeters), "m" (meter), and "cm" (centimeters). ApplyNetFit will convert to cm.

relPower
The fishing power of each net size. Leave NULL (default) if only single mesh or all meshes are fished equally. Otherwise, supply numeric vector (length = number of mesh sizes) representing net configuration effort. For example, if net mesh sizes are 6", 7", & 8" and the 8" mesh is fished at twice the effort, relPower = c(1, 1, 2) [for 6", 7", & 8"].

Details

ApplyNetFit fits 5 Millar models: norm.loc; norm.sca; lognorm; binorm.sca; and bilognorm. Each one of these models is a list with 12 items (explained below).

Value

A list of model fits (see details).

Value

ApplyNetFit returns a list of class "NetFit" each list item containing the components below.

par see **Value** in optim

value see **Value** in optim

counts see **Value** in optim

convergence see **Value** in optim

message see **Value** in optim

hessian see **Value** in optim

Deviance model deviance (value)

deviance model deviance (function); not sure of purpose

rtype the name (as character) of the model

rel.power values supplied to relPower, relative fishing power of the gear

Meshsize mesh size (in centimeters)

Data length frequency by mesh size (raw count data)

References

https://www.stat.auckland.ac.nz/~millar/selectware/RNext/

https://www.stat.auckland.ac.nz/people/rmil013

Millar and Holst (1997). Estimation of gillnet and hook selectivity using log-linear models ICES J. Mar. Sci. 54: 471-477.

Examples


ApplyNetFit(
  data = trammel_catch,
  len = FL,
  mesh = MeshSize,
  meshUnit = "in",
  relPower = c(1, 1, 2)
)

[Package spopmodel version 0.0.0.9000 Index]