BackCalcLength {spopmodel}	R Documentation

Uses Dahl-Lea method to back-calculate length at age.

Description

BackCalcLength employs the Dahl-Lea algorithm to back-calculate length at age given specific data from the analysis of hard parts (e.g., scales). Assumes for now first age is 1.

Usage


BackCalcLength(data, lCap, rCap, rAtAge)

Arguments

data
Dataframe where each row is a single fish & contains fields below.

lCap
Field name in data containing (numeric) length at capture.

rCap
Field name in data containing (numeric) scale radius at capture.

rAtAge
Field (column) names or column numbers (easier) in data containing scale radius at each age i.

Value

A dataframe with two numeric fields: len & age. NA values have been removed.

References

Lea, E. 1910. On the methods used in the herring-investigations. Publ. Circonst. Cons. perm. int. Explor. Mer 108(1):14–22.

Ogle, D.H. 2016. Introductory Fisheries Analyses with R. Chapman & Hall/CRC, Boca Raton, FL.

Examples


# code here

[Package spopmodel version 0.0.0.9000 Index]