ChapmanRobson {spopmodel}	R Documentation

Calculates survival rate given age frequency.

Description

ChapmanRobson adapted from Ricker 1975, eqn. 2.5 (2.6 for variance). Method is similar to using catch curve with age frequency.

Usage


ChapmanRobson(ages, n, N0 = 2)

Arguments

ages
Numeric vector of ages.

n
Numeric vector (length of ages) containing count (n) at each age.

N0
Single value denoting starting age.

Value

Survival rate estimate \hat{S} with variance Var(\hat{S}). Also returns sumages used * count and sumfrequency of all ages used

References

Ricker, W.E. (1975). Computation and Interpretation of Biological Statistics of Fish Populations. Bulletin of the Fisheries Research Board of Canada, Bulletin 191, Ottawa.

Examples


# ChapmanRobson(ages = , n = , N0 = 3)

[Package spopmodel version 0.0.0.9000 Index]