DeviancePlots {spopmodel}	R Documentation

Generate Millar's deviance plots for gear selectivity.

Description

We choose a selectivity model based on deviance. The lower the deviance, the better the fit. Deviance plots display residuals (by mesh size) in a "bubble-plot" format. (describe 'bubbles' and size here) Plot is a side effect of Millar's Summary function. Function returns a list, with number of items equal to the number of models (i.e., 5). See Values for output.

Default S3 method.

S3 method for class NetFit.

Usage


DeviancePlots(x, ...)

## Default S3 method:
DeviancePlots(x, ...)

## S3 method for class 'NetFit'
DeviancePlots(x, ...)

Arguments

x
Generic parameter for S3 method. Likely a model fit of class "NetFit".

...
Passed to other methods.

Value

If x is a list, output is a list with same number of elements, each element is a matrix. Otherwise output is a matrix with one column. See Values.

Values

null.l

doc needed here

model.l

doc needed here

full.l

doc needed here

Deviance

model deviance

Pearson.chisq

doc needed here

d.o.f

degrees of freedom

Note

DeviancePlots is essentially a convenient wrapper for Millar's Summary function.

Examples


# for method .NetFit
net_fit <- ApplyNetFit(
  data = trammel_catch,
  len = FL,
  mesh = MeshSize,
  meshUnit = "in",
  relPower = c(1,1,2)
)
DeviancePlots(net_fit)

[Package spopmodel version 0.0.0.9000 Index]