Exploitation {spopmodel}	R Documentation

Calculcates adjusted exploitation rate μ\prime.

Description

Adjusted exploitation compensates for lack of angler- reporting (i.e., via reporting rate), tag loss, and mortality due to tagging. If tag loss or tagging mortality are unknown, values from literature can be used.

μ\prime=\frac{μ}{(repRate \times (1-tagL) \times (1-tagM))}

Usage


Exploitation(rr, rt, repRate, tagL, tagM)

Arguments

rr
Number of non-reward tags reported (positive integer).

rt
Number of non-reward tags released (positive integer).

repRate
Reporting rate RepRate from ReportingRate.

tagL
Fraction of tags lost (default ??).

tagM
Mortality (as a fraction) from tagging (default ??).

Value

List containing number returned, number released (tagged), reporting rate, raw exploitation rate mu, and adjusted exploitation rate μ\prime.

References

Kevin A. Meyer & Daniel J. Schill (2014) Use of a Statewide Angler Tag Reporting System to Estimate Rates of Exploitation and Total Mortality for Idaho Sport Fisheries, North American Journal of Fisheries Management, 34:6, 1145-1158, DOI: 10.1080/02755947.2014.951803

Examples


# numbers for example only
Exploitation(rr = 5, rt = 175, repRate = 0.5, tagL = 0.05, tagM = 0.05)

[Package spopmodel version 0.0.0.9000 Index]