FitVBGM {spopmodel}	R Documentation

Fits von Bertalanffy growth model (VBGM) using stats::nls.

Description

FitVBGM using Ricker 1975, eqn. 9.9 to calculate length at age.l_t = L_{inf}(1 - e^{-K(t-t_0)})

Usage


FitVBGM(data, len, age, start = list(), ...)

Arguments

data
Dataframe with length and age fields.

len
Numeric length field within data.

age
Numeric age field within data.

start
Starting parameters passed to stats::nls. Must be a named list with values for "Linf", "K", and "t0".

...
Further arguments passed to nls.

Value

Model fit object of type nls.

References

Ricker, W.E. (1975). Computation and Interpretation of Biological Statistics of Fish Populations. Bulletin of the Fisheries Research Board of Canada, Bulletin 191, Ottawa.

Examples


# start_vals from Cal Fish and Game, Kohlhorst et al 1980
start_vals <- list(Linf = 261.2, K = 0.04027, t0 = 3.638)
catch_data <- subset(trammel_catch, !is.na(Age))
# below example giving error...need to troubleshoot
## Not run: 
FitVBGM(
  data = subset(trammel_catch, !is.na(Age)),
  len = FL,
  age = Age,
  start = start_vals
)

## End(Not run)

[Package spopmodel version 0.0.0.9000 Index]