Frequency {spopmodel}	R Documentation

Create a frequency distribution of x by binWidth.

Description

Frequency is a custom wrapper (mostly) for hist. It simplifies input to only data and bin width. Bins are always left included but not right, as in `[a, b)`. Frequency removes NA values (with warning) prior to binning.

Usage


Frequency(x, binWidth)

## S3 method for class 'Frequency'
summary(object, ...)

## S3 method for class 'Frequency'
print(x, ...)

Arguments

x
A postive numeric vector to be binned according to binWidth.

binWidth
A numeric scalar (integer) defining the bin size.

object
Object of class Frequency.

...
Passed on to other methods.

Value

A list (of class Frequency & class hist) containing proportion & count frequencies.

Methods (by generic)

summary: summary method (See notes).

print: print method (See notes).

Values

others

see hist

.

xna

number of NA values removed

.

xname

deparsed name of x, replaces hist xname

binw

value passed to binWidth

xstats

function when run will yield descriptive statistics of x

Note

Method summary written for class Frequency. Returns dataframe with fields Bin, Count, Density, Proportion. Print method also writted for class Frequency.

Examples


Frequency(trammel_catch[["FL"]], binWidth = 5)

[Package spopmodel version 0.0.0.9000 Index]