LoadMillarCode {spopmodel}	R Documentation

Download R. Millar gear selectivity code.

Description

LoadMillarCode downloads to user's computer gear selectivity source code written by Russel Millar. Files downloaded currently are NextGeneration.R & SelnCurveDefinitions.R, the main two used in this package. Files download to root directory.

Usage


LoadMillarCode()

Value

Message giving file path to where files were downloaded.

Note

Only run this code if interested in seeing source files. Other- wise, gear selectvity functions within this package will suffice.

References

https://www.stat.auckland.ac.nz/~millar/selectware/RNext/

https://www.stat.auckland.ac.nz/people/rmil013

Millar and Holst (1997). Estimation of gillnet and hook selectivity using log-linear models ICES J. Mar. Sci. 54: 471-477.

Examples


## Not run: 
LoadMillarCode()

## End(Not run)

[Package spopmodel version 0.0.0.9000 Index]