PopProjections {spopmodel}	R Documentation

Get population projection over defined period (years).

Description

PopProjections is essentially a wrapper for pop.projection. It will run pop.projection for every simulation (i.e., column) in the simulation matrices (passed to fSims & sSims). Function creates the Leslie Matrix (*A*) required by pop.projection.

Usage


PopProjections(fSims, sSims, mn, sdev, ageFreq, period)

Arguments

fSims
A matrix of fecundity simulations, likely the product of simulated number of eggs x simulated spawning probabilities x sex ratio (expressed as a proportion, e.g., 0.5).

sSims
A matrix of survival simulations, likely one matrix for each level of exploitation (mu) being tested.

mn
Scalar numeric value as survival probability of oldest fish (i.e., max age in ageFreq)

sdev
Scalar numeric value as corresponding error for mn.

ageFreq
Numeric vector of age frequency passed on to n in pop.projection.

period
Scalar integer given number of years to project. Passed on to iterations in pop.projection.

Value

A large matrix with rows as 'Value' output in link{pop.projection} and number of columns = number of simulations.

Examples


# code here

[Package spopmodel version 0.0.0.9000 Index]