RelativeRetention {spopmodel}	R Documentation

Get gear relative retention based on net fit model output.

Description

RelativeRetention is a re-write of Millar's PlotCurves for convenient use within this package. It generates (among other items) a dataframe relative retention values for each gear type (columns) by each length (or length bin; rows). Out put can then be used to 'adjust' catch accordingly.

Default S3 method.

S3 method for class NetFit.

Usage


RelativeRetention(x, ...)

## Default S3 method:
RelativeRetention(x, ..., standardize = TRUE)

## S3 method for class 'NetFit'
RelativeRetention(x, ...)

Arguments

x
Generic for S3 methods. Model output from NetFit.

...
Passed to other methods.

standardize
Boolean. If TRUE (default), relative retention values are divided by max(relative retention). If FALSE, not such division occurs.

Value

List of objects. See Values.

Values

Data

dataframe of relative retention values

Freq

raw frequency data (i.e., recorded catch)

RelPower

relative power of gear type

MeshSize

mesh size of each gear

Deviance

measure of 'fit' for each model

ColNames

column names of Data

Model

name of net fit model

Standardized

true or false, is relative retention standardized

Examples


# for .NetFit method
net_fit <- ApplyNetFit(
  data = trammel_catch,
  len = FL,
  mesh = MeshSize,
  meshUnit = "in",
  relPower = c(1,1,2)
)
RelativeRetention(net_fit)

[Package spopmodel version 0.0.0.9000 Index]