ReportingRate {spopmodel}	R Documentation

Calculates angler reporting rate of non-reward tags.

Description

In mark-recapture studies, anglers are encouraged to return tags to the issuing entity. A mixture of non-reward tags and high-value reward tags typically are released to ensure returns. Not all anglers return tags (non-reporting), and the reporting rate adjusts for non-reporting.

Usage


ReportingRate(rr, rt, nr, nt)

Arguments

rr
Number of non-reward tags reported (positive integer).

rt
Number of non-reward tags released (positive integer).

nr
Number of high-reward tags reported (positive integer).

nt
Number of high-reward tags released (positive integer).

Value

List containing reporting rate and reporting rate variance.

Note

In the case of CDFW sturgeon mark-recapture, all tags released are reward tags. We assume, then, a reporting rate of 100 the highest value reward.

References

Kevin A. Meyer & Daniel J. Schill (2014) Use of a Statewide Angler Tag Reporting System to Estimate Rates of Exploitation and Total Mortality for Idaho Sport Fisheries, North American Journal of Fisheries Management, 34:6, 1145-1158, DOI: 10.1080/02755947.2014.951803

Examples


# numbers for example only
ReportingRate(rr = 5, rt = 175, nr = 10, nt = 175)

[Package spopmodel version 0.0.0.9000 Index]