fin_ray_aging {spopmodel}	R Documentation

Fin Ray Analysis of San Francisco Estuary-based White Sturgeon.

Description

A dataset with 100 observations & variables lencap, radcap, and anu<number>, where number is 1–>10. These data represent required format for using BackCalcLength(). Data are from White Sturgeon fin ray sampling & ageing.

Format

A data frame with 100 observations and 12 variables:

lencap

length at capture (in this case fork length)

radcap

radius at capture

anu1

annulus at age-1

anu2

annulus at age-2

anu3

annulus at age-3

anu4

annulus at age-4

anu5

annulus at age-5

anu6

annulus at age-6

anu7

annulus at age-7

anu8

annulus at age-8

anu9

annulus at age-9

anu10

annulus at age-10

Details

These data establish back-calculated length at age. Data are real but subsetted here for ease of demonstration.

Source

White Sturgeon fin rays subset 2014 sampling (CDFW; Univ. of Idaho).

[Package spopmodel version 0.0.0.9000 Index]