input {spopmodel}	R Documentation

Suite of functions to generate model data inputs.

Description

spopmodel requires specific inputs to operate. In the absence of such inputs, functions herein can generate dataframes to satisfy model requirements. If such data already exists, however, then it's best to use those data. NOTE: Functions herein default for use with San Francisco Estuary-based White Sturgeon. So do keep that in mind when employing these functions.

Uses glm to predict spawning probability based on probability of maturity at length i. Spawning probability is then adjusted for fraction of females spawning annually.

Uses lm to predict number of eggs given length i.

Uses age frequency to provide by age (1) estimated abundance & (2) estimated abundance of females. Uses arguments supplied to abund & fracFemale to achieve this.

Usage


SpawningProb(pMat = NULL, len, age, mature = 0.15)

EggCount(numEggs = NULL, len, age)

AgeDist(ageFreq, abund = 48000, fracFemale = 0.5)

SurvivalProb(..., mu = 0.01, agesMu = 10:15, estS = 0.8132,
  estMu = 0.1364, methodSB = TRUE)

CreateSPDataframe(ages, sRate, sRateErr)

Arguments

pMat
A dataframe with two fields: length & probability of maturity for each length. Default uses internal SpawningProbWST (for White Sturgeon) & from Chapman 1989 (?).

len
Numeric vector of lengths supplied as new data to predict.

age
Numeric vector of ages, where each age is appropriate for given length supplied in len. Vector must be same length as len.

mature
Fraction of females spawning in a given year. Default 0.15.

numEggs
A dataframe with two fields: length & number of eggs. for each length. Default uses internal FecundityWST (for White Sturgeon) & from Devore 19??, using 0.072 * len^2.94.

ageFreq
A numeric vector of age frequency (like that derived from age-length key or direct ageing).

abund
A numeric scalar given estimated overall abundance for species being modeled. Default 48,000 (as estimated by CDFW).

fracFemale
A numerice scalar given fraction of females in population. Default 0.5 from Chapman et al. 1996.

...
Passed to internal function CreateSPDataframe.

mu
A numeric vector giving exploitation (harvest rate; between 0 and 1). Default 0.01.

agesMu
A numeric vector of ages on which exploitation applies (i.e., age group subject to harvest). Default 10:15.

estS
A numeric scalar given estimated survival rate of (typically) adult fish. Derived using ChapmanRobson. Default 0.8132 from S. Blackburn thesis work.

estMu
A numeric scalar given estimated exploitation rate of harvestable fish. Derived using Exploitation. Default 0.1364 from S. Blackburn thesis work using 2014-2016 CDFW mark-recapture data.

methodSB
Logical. Default (TRUE) denotes using S. Blackburn method to calculate F using Z & A calculated from estS and not from S0. Change to FALSE to use S0.

ages
A numeric vector of ages (e.g., 0:19).

sRate
A numeric vector giving survival rate (between 0 and 1).

sRateErr
A numeric vector given standard error of sRate.

SpawningProb

NA

EggCount

NA

AgeDist

NA

SurvivalProb

NA

CreateSPDataframe

NA

Note

Ideally ages vector supplied from age distribution data (e.g., age_dist). Vector should contain age-0 sequentially through oldest age. sRate and sRateErr need not match ages in length, just not longer. If less than, the last value (e.g., sRate[length(sRate)] is repeated to complete the dataframe. sRate[length(sRate)] must be S0, that is survival rate considering only natural mortality.

[Package spopmodel version 0.0.0.9000 Index]