number_eggs {spopmodel}	R Documentation

Number of eggs by age for San Francisco Estuary-based White Sturgeon.

Description

A dataset for White Sturgeon (Acipenser transmontanus) with 20 ages (0-19) and corresponding egg count with associated error.

Format

A data frame with 20 rows and 3 variables:

age

numeric age from 0-19

count

number of eggs at each age

se

error associated with count

Details

Based on linear regression using data from DeVore et al. 1995. DeVore equation 0.072 * len^{2.94} to get number of eggs.

Source

Model development input data (CDFW; Univ. of Idaho).

References

DeVore, J. D., B. W. James, C. A. Tracy, and D. A. Hale. 1995. Dynamics and potential production of White Sturgeon in the Columbia River downstream from Bonneville Dam. Transactions of the American Fisheries Society 124:845–856.

[Package spopmodel version 0.0.0.9000 Index]