plot.Frequency {spopmodel}	R Documentation

Generate base R histogram from object of class Frequency.

Description

Plot a histogram using hist (esentially calling next method on plot, which calls plot.histogram). Display contains sample size along with some descriptive statistics.

Usage


## S3 method for class 'Frequency'
plot(x, ...)

Arguments

x
An object of class Frequency (& histogram).

...
Passed on to other methods.

Value

Displays a histogram using frequency item in x.

Examples


freq <- Frequency(trammel_catch[["FL"]], binWidth = 5)
plot(freq)

[Package spopmodel version 0.0.0.9000 Index]