plot.RelativeRetention {spopmodel}	R Documentation

Plot relative retention values for each gear size.

Description

This is an S3 method for plot of class RelativeRetention. It allows the user to visually display relative retentions of each gear size and to view on each plot model deviance (for ease of deciding best fit).

Usage


## S3 method for class 'RelativeRetention'
plot(x, ...)

Arguments

x
Object of class RelativeRetention.

...
Passed on to other methods.

Value

Nothing. But does display (base R) line plot of relative renetions as a function of measured length (e.g., fork length). Plot title contains model name, model deviance, and true or false for "is standardized?". See default method RelativeRetention.

Examples


net_fit <- ApplyNetFit(
  data = trammel_catch,
  len = FL,
  mesh = MeshSize,
  meshUnit = "in",
  relPower = c(1,1,2)
)
rr <- RelativeRetention(net_fit)
plot(rr$norm.loc)

[Package spopmodel version 0.0.0.9000 Index]