prob_spawn {spopmodel}	R Documentation

Spawning probability by age for San Francisco Estuary-based White Sturgeon.

Description

A dataset for White Sturgeon (Acipenser transmontanus) with 20 ages (0-19) and corresponding spawning probability with associated error.

Format

A data frame with 20 rows and 3 variables:

age

numeric age from 0-19

prob

probability of spawning at each age

se

error associated with prob

Details

Based on generalized linear model using data from Chapman 1989. Model used binomial(link = "logit") with probability of maturity ~ length.

Source

Model development input data (CDFW; Univ. of Idaho).

References

Chapman, F. 1989. Sexual maturation and reproductive parameters of wild and domestic stocks of White Sturgeon, Acipenser transmontanus. Doctoral dissertation. University of California, Davis, California.

[Package spopmodel version 0.0.0.9000 Index]