spopmodel {spopmodel}	R Documentation

spopmodel: A package for modeling fish population dynamics.

Description

spopmodel employs a female-based Leslie Matrix to evaluate population dynamics (specifically population growth rate, λ) over varying exploitation (μ). Model inputs expects age-specific probalities of spawning & survival, age-specific fecundity, and age distribution proportional to some estimated abundance. spopmodel's creation stems from the collaborative efforts of University of Idaho and CA Department of Fish & Wildlife.

Age Distribution

Age frequency scaled to some absolute abundance estimate (see example below). Model requires young (i.e., age-0; age-1) fish, so values from literature (or other sources) may be necessary to complete data gaps.

Example:

freq_{age_{i}} \times \hat{abund}

Survival Probability

Survival at age (as assigned in Age Distribution). Again values can be filled in where needed. Probability to be adjusted downward for ages affected by exploitation (μ). In the case of White Sturgeon for example, survival estimated at ~0.94, and for ages 10-15 adjusted downard by ~0.13. These ages vunerable to fishing mortality.

Spawning Probability

Probability based on first age of maturity. For White Sturgeon (as an example), this is ~age-10. Again, probability should be in-line with ages set in Age Distribution, and this may require setting to 0 age-0 through age-<first age maturity>-1.

Fecundity

Number of eggs per age, starting with first age of maturity. If data are lacking, may need to consult literature for species in question.

Note

For probability datasets & fecundity, some measure of error (i.e., standard deviation) per value is required. These values are used in the simulation process.

[Package spopmodel version 0.0.0.9000 Index]