trammel_catch {spopmodel}	R Documentation

Catch data of San Francisco Estuary-based White Sturgeon.

Description

A dataset with 1000 observations and 3 variables. Data are from CA Department of Fish & Wildlife (CDFW) sturgeon mark- recapture survey. Trammel nets of varying mesh size are deployed in bay waters during late Summer and early Fall. Data are specific for White Sturgeon (Acipenser transmontanus).

Format

A data frame with 1000 rows and 3 variables:

MeshSize

trammel net mesh size in inches: 6 or 7 or 8

FL

measured fork length in centimeters

Age

age based on fin ray analysis

Details

Data are from 2014-2016 sampling. Some fenagling of data was necessary for use within this package.

Source

CDFW sturgeon mark-recapture survey https://www.wildlife.ca.gov/Conservation/Delta/Sturgeon-Study

[Package spopmodel version 0.0.0.9000 Index]